Homework - lesson2 - level2 done
