+++ src/main/java/students/student_svetlana_seda/lesson2/level2/HelloLana.java
@@ -1,0 +1,16 @@
+package students.student_svetlana_seda.lesson2.level2;
+
+import java.util.Scanner;
+
+public class HelloLana {
+    public static void main(String[] args) {
+
+            String n;
+            Scanner scanner = new Scanner(System.in);
+            System.out.println("Please, enter your username: ");
+            n = scanner.nextLine();
+                 System.out.println("HELLO, " + n + "!");
+
+    }
+
+}
