--- src/main/java/students/student_svetlana_seda/lesson2/level2/HelloWorld.java
@@ -1,10 +1,0 @@
-package students.student_svetlana_seda.lesson2.level2;
-
-public class HelloWorld {
-    public static void main(String[] args) {
-
-        System.out.println("Hello World!");
-
-    }
-
-}
