--- src/main/java/students/student_svetlana_seda/lesson2/level2/Task8.java
@@ -1,19 +1,0 @@
-package students.student_svetlana_seda.lesson2.level2;
-
-import org.w3c.dom.ls.LSOutput;
-
-import java.util.Scanner;
-
-public class Task8 {
-    public static void main(String[] args) {
-
-    String n;
-    Scanner scanner = new Scanner(System.in);
-        System.out.println("Please, enter your username: ");
-          n = scanner.nextLine();
-
-        System.out.println("HELLO, " + n + "!");
-
-    }
-
-}
